PPO 多行程 rollout 2.21 -> 0.38 秒/局，並從監督式 checkpoint 熱啟動
rollout 的 88% 是引擎和 gen0（純 CPU），網路只佔 12%，所以 10 個 worker 幾乎
線性加速。spatial 存 float16 讓軌跡減半，前向也吃還原後的值，所以 logprob
往返誤差還是 1.526e-05。update 放 GPU 快 5.8 倍。

隨機初始整季 0 元、每步 1.17 個 unit；熱啟動 ckpt-e2e-round6（encoder_version
5，60 個張量全吃得下）是 25,000 元、9.84 個 unit。第一次真跑進行中。

細節見 journal 2026-08-28 §34。

+++ model/ppo.py
@@ -196,7 +196,7 @@ class Trajectory:
     期末。reward 由 `step_rewards` 從它算出來，不另外存。
     """
 
-    spatial: np.ndarray        # [T, N_SPATIAL, 10, 10] float32
+    spatial: np.ndarray        # [T, N_SPATIAL, 10, 10] float16（見 half_obs）
     scalar: np.ndarray         # [T, N_SCALAR] float32
     value: np.ndarray          # [T] float32
     logp: np.ndarray           # [T] float32（unit 加 market，整步的和）
@@ -330,7 +330,9 @@ class RolloutBatch:
             return torch.as_tensor(a, device=device)
 
         return {
-            "spatial": t(self.spatial[idx]),
+            # rollout 存的是 float16（見 `VecRollout.half_obs`）；還原成
+            # float32 之後跟 rollout 前向吃的位元完全一樣。
+            "spatial": t(self.spatial[idx].astype(np.float32, copy=False)),
             "scalar": t(self.scalar[idx]),
             "unit_board": t(ub),
             "unit_pos": t(self.unit_pos[u]),
